+++ CHANGELOG.rst
@@ -2,6 +2,9 @@ Changelog
 =========
 
 
+- Rename ``Cache.setup()`` to ``Cache.configure()``. **breaking change**
+
+
 v0.1.0 (2018-01-28)
 -------------------